simpler implementation for homepage customization
implemented in geonode/theme
