--- geonode/client/templates/admin/themes/custom_theme_base_header.txt
@@ -1,4 +1,0 @@
-{% extends "base.html" %}
-{% load i18n %}
-{% load staticfiles %}
-{% load base_tags %}
