--- geonode/client/templates/admin/themes/custom_theme_bigsearch_footer.txt
@@ -1,1 +1,0 @@
-{% endblock bigsearch %}
