--- geonode/client/templates/admin/themes/custom_theme_bigsearch_header.txt
@@ -1,1 +1,0 @@
-{% block bigsearch %}
