--- geonode/client/templates/admin/themes/custom_theme_contactus_header.txt
@@ -1,2 +1,0 @@
-{% block contactus %}
-{{ block.super }}
