--- geonode/client/templates/admin/themes/custom_theme_footer_header.txt
@@ -1,2 +1,0 @@
-{% block footer %}
-{{ block.super }}
