--- geonode/client/templates/admin/themes/custom_theme_hero_footer.txt
@@ -1,1 +1,0 @@
-{% endblock hero %}
