--- geonode/client/templates/admin/themes/custom_theme_hero_header.txt
@@ -1,1 +1,0 @@
-{% block hero %}
