--- geonode/client/templates/admin/themes/custom_theme_partners_footer.txt
@@ -1,2 +1,0 @@
-    {% endblock partners %}
-{% endblock extra_mainbody %}
