--- geonode/client/templates/admin/themes/custom_theme_partners_header.txt
@@ -1,2 +1,0 @@
-{% block extra_mainbody %}
-    {% block partners %}
