+++ geonode/static/geonode/css/base.css
@@ -7016,6 +7016,7 @@ article.tab-pane {
   padding-bottom: 0;
 }
 .home .jumbotron {
+  position: relative;
   background: #2c689c;
   color: white;
   margin-bottom: 0;
@@ -7129,6 +7130,12 @@ article.tab-pane {
 }
 .home #showcase p {
   font-size: 18px;
+}
+.home .partner_logo {
+  background-repeat: no-repeat;
+  background-position: center;
+  background-size: contain;
+  height: 250px;
 }
 /* autocomplete
 ------------------------*/
@@ -7655,6 +7662,20 @@ footer a:hover {
 }
 footer form {
   color: #333333;
+}
+/* Copyright
+----------------------- */
+.footer-copyright {
+  padding: 30px 0;
+}
+.footer-copyright .copyright {
+  color: #fff;
+  text-align: center;
+  text-transform: uppercase;
+}
+.footer-copyright .copyright small {
+  text-transform: uppercase;
+  font-size: 14px;
 }
 /* Etc
 ----------------------- */
