+++ geonode/static/geonode/less/base.less
@@ -181,6 +181,7 @@ article.tab-pane {
 	  padding-bottom: 0;
 	}
 	.jumbotron {
+    position: relative;
 		background: @brand-primary;
 		color: white;
 		margin-bottom: 0;
@@ -281,7 +282,13 @@ article.tab-pane {
 			}
 		}
 		p {font-size: 18px;}
-	}
+  }
+  .partner_logo {
+    background-repeat: no-repeat;
+    background-position: center;
+    background-size: contain;
+    height: 250px;
+  }
 }
 /* autocomplete
 ------------------------*/
@@ -838,6 +845,21 @@ footer {
   }
   form {
 	  color: @text-color;
+  }
+}
+
+/* Copyright
+----------------------- */
+.footer-copyright {
+  padding: 30px 0;
+  .copyright {
+    color: #fff;
+    text-align: center;
+    text-transform: uppercase;
+  }
+  .copyright small {
+    text-transform: uppercase;
+    font-size: 14px;
   }
 }
