+++ geonode/templates/index.html
@@ -16,10 +16,9 @@
   {% else %}
   <div class="jumbotron">
   	<div class="container">
-  		<h1>{% trans "Welcome" %}</h1>
+        <h1>{{custom_theme.jumbotron_welcome_title|default:_("Welcome")}}</h1>
   		<p></p>
-  		<p>{% blocktrans %}GeoNode is an open source platform for sharing geospatial data and maps.{% endblocktrans %}</p>
-		<p>{% blocktrans %}If you have any questions about the software or service, join our <a href="http://lists.osgeo.org/cgi-bin/mailman/listinfo/geonode-users">mailing list</a>.{% endblocktrans %}</p>
+        <p>{{custom_theme.jumbotron_welcome_content|default:_("GeoNode is an open source platform for sharing geospatial data and maps.")}}</p>
   		<a class="btn btn-default btn-lg" target="_blank" href="http://docs.geonode.org/en/master/tutorials/users/index.html" role="button">{% blocktrans %}Get Started &raquo;{% endblocktrans %}</a>
   	</div>
   </div>
@@ -33,6 +32,11 @@
       {{ block.super }}
       {% else %}
       <div class="big-search">
+          {% if custom_theme.jumbotron_site_description %}
+          <div class="container">
+            <p>{{ custom_theme.jumbotron_site_description }}</p>
+          </div>
+          {% endif %}
       	<div class="container">
       		<h1>{% trans "Search for Data." %}</h1>
       		<div class="search">
@@ -310,6 +314,23 @@
         </div>
       </section>
       {% endblock showcase %}
+      {% block partners %} 
+        {% if custom_theme.partners_title %}
+        <section id="partners">
+        <div class="container text-center">
+          <h2>{{ custom_theme.partners_title }}</h2>
+          <div class="row">
+              {% for partner in custom_theme.partners.all %}
+              <div class="col-md-6">
+                  <div class="partner_logo"{% if partner.logo %} style="background-image: url({% if %}{{partner.logo.url}}{% else %}{% static 'geonode/img/missing_thumb.png' %}{% endif %})"{% endif %}></div>
+                  <div><a href="{{ partner.partner_link }}" target="_blank">{{ partner.title }}</a></div>
+              </div>
+              {% endfor %}
+          </div>
+        </div>
+        </section>
+        {% endif %}
+      {% endblock partners %} 
 
   {% endblock mainbody %}
